api/requirements.txt
Django==2.1.5
django-cors-headers==2.4.0
django-filter==2.0.0
djangorestframework==3.9.0
mysql-connector==2.1.6
mysqlclient==1.3.14
pytz==2018.7
